fixed displayCommment function according to the functional requirement of BandSite sprint-2

+++ scripts/index-page.js
@@ -1,18 +1,8 @@
-
 const link = "https://project-1-api.herokuapp.com/comments";
 const apiKey = "?api_key=3e76ec52-db2a-4564-99d4-11b55158c9e7";
 
-
-function displayComments() {
-  const getComments = axios.get(link + apiKey);
-  getComments.then((res) => {
-    const commentsData = res.data;
-    const commentsArrey = commentsData.sort(
-      (a, b) => b.timestamp - a.timestamp
-    );
-    commentsList.innerHTML = "";
-    commentsArrey.forEach((comment) => {
-      const commentLi = document.createElement("li");
+function displayComment(comment){
+        const commentLi = document.createElement("li");
       commentLi.classList.add("item");
       commentsList.appendChild(commentLi);
 
@@ -56,34 +46,49 @@ function displayComments() {
       const likeImgLink = document.createElement("a");
       likeImgLink.classList.add("item__actions", "item__actions--like");
       likeImgLink.setAttribute("data-id", comment.id);
-      likeImgLink.setAttribute("onclick", "liked(this)");
+      // likeImgLink.setAttribute("onclick", "liked(this)");
+      likeImgLink.setAttribute("value", "like");
       itemFooter.appendChild(likeImgLink);
 
       const likeImg = document.createElement("img");
       likeImg.classList.add("like-img");
       likeImg.setAttribute("src", "./assets/Icons/SVG/icon-like.svg");
       likeImg.setAttribute("data-id", comment.id);
+      likeImg.setAttribute("value", "like");
       likeImgLink.appendChild(likeImg);
 
       const likeCounter = document.createElement("p");
       likeImg.classList.add("like-counter");
       likeCounter.innerText = comment.likes;
-      likeCounter.setAttribute("data-id", comment.id);
       likeImgLink.appendChild(likeCounter);
 
       const deleteLink = document.createElement("a");
       deleteLink.classList.add("item__actions", "item__actions--delete");
       deleteLink.setAttribute("data-id", comment.id);
       deleteLink.innerText = "Delete";
-      deleteLink.setAttribute("onclick", "deleteComment(this)");
+      // deleteLink.setAttribute("onclick", "deleteComment(this)");
+      deleteLink.setAttribute("value", "delete");
       itemFooter.appendChild(deleteLink);
+}
+
+function showComments() {
+  const getComments = axios.get(link + apiKey);
+  getComments.then((res) => {
+    const commentsData = res.data;
+    const commentsArrey = commentsData.sort(
+      (a, b) => b.timestamp - a.timestamp
+    );
+    commentsList.innerHTML = "";
+    commentsArrey.forEach((comment) => {
+      displayComment(comment);
+
     });
   });
 }
 
 const form = document.querySelector(".form");
 const commentsList = document.querySelector(".comments__list");
-displayComments();
+showComments();
 
 form.addEventListener("submit", (event) => {
   event.preventDefault();
@@ -96,22 +101,27 @@ form.addEventListener("submit", (event) => {
   const addComment = axios.post(link + apiKey, postedComment);
   addComment.then((res) => {
     event.target.reset();
-    displayComments();
+    showComments();
   });
 });
 
-function liked(d){
-  const userId = d.getAttribute("data-id");
-  const addLike = axios.put(link+"/"+ userId+"/like"+apiKey)
-  addLike.then((res)=>{
-    displayComments()
-  })
-}
 
-function deleteComment(d){
-  const userId = d.getAttribute("data-id");
-  const addLike = axios.delete(link+"/"+ userId+apiKey)
-  addLike.then((res)=>{
-    displayComments()
-  })
-}
+const itemActions = document.querySelector(".comments__list");
+
+itemActions.addEventListener("click", (event) => {
+
+  if (event.target.getAttribute("data-id")) {
+    const userId = event.target.getAttribute("data-id");
+    if (event.target.getAttribute("value") === "like") {
+      const addLike = axios.put(link + "/" + userId + "/like" + apiKey);
+      addLike.then((res) => {
+        showComments();
+      });
+    } else {
+      const delComment = axios.delete(link + "/" + userId + apiKey);
+      delComment.then((res) => {
+        displayComments();
+      });
+    }
+  }
+});
